+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## X.X.X
 * Added support for server requested SDK log gathering.
+* Added support for SDK behavior settings that control the SDK's automatic session tracking, automatic view tracking, automatic crash reporting, and Journey Trigger Views.
 
 ## 26.1.4
 * Added support for reporting the app's current theme (light or dark) when presenting feedback widgets, rating widgets, and content, so they are displayed in matching conditions.